+++ ESP32/console_esp32.h
@@ -21,8 +21,8 @@ private:
 	int               led_pulses = 0;
 
 protected:
-	int wait_for_char_ll(const short timeout) override;
-	void put_char_ll    (const char c) override;
+	int  wait_for_char_ll(const int  timeout) override;
+	void put_char_ll     (const char c      ) override;
 
 public:
 	console_esp32(kek_event_t *const stop_event, comm *const io_port, const int t_width, const int t_height);
